[#907] Add interruptible to pipelines
Per https://docs.gitlab.com/ee/ci/yaml/#interruptible, pipelines can be
automatically stopped if a newer pipeline is started. This commit takes care
of this change. This is done by adding it to the shared configuration
section `.build` and a new one `.docs` for shared documentation
configuration. Since the commit verify job does not use either one of
these, it has the `interruptible` set directly on it.

Also, updated commit-verify to now use Ubuntu 22.04 image.