[#885] MariaDB testing support + Documentation
In examining whether we can fully support MariaDB instead of MySQL
(which is what issue #885 wants), we found out that we exactly emulate
MySQL to the level of being bug by bug comptabile in the DATE_FORMAT()
function; but MariaDB changed how the function is implemented, resulting
in a different set of bugs. Discussion over here:
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/885#note_1107901501. As
a result, we disabled the tests that fail if we are running on MariaDB
in `tests/test_date_time_functions.bats.in`.

Documentation for installing/configuring MariaDB, as well as
documentation telling users that MariaDB is equivalent to MySQL, has
been updated.

- `Dockerfile-Tests.rocky`, `Dockerfile-Tests.ubuntu` have instructions
  for running the MariaDB docker image.
- `doc/developer_doc.rst` mentions the DNF/APT packages that need to be
  installed.
- `doc/config.rst`, `doc/intro.rst`, and `doc/rocto.rst` updated to
  mention MariaDB as equivalent to MySQL.