[#908] Only create/push docker images after merges + small fix
We had an issue were we created docker images during scheduled jobs when
we didn't need to, as technically nothing changed since the previous
merge. This commit uses $CI_PIPELINE_SOURCE to detect if we are
"scheduled" or a "push", and only builds/pushes docker images on a push
(which is a fast-forward merge for us, as we don't do merge commits).
$CI_COMMIT_BRACH and $CI_PROJECT_PATH are both used to ensure that this
happens on the master branch of the main Octo repo only.

Yesterday's change in commit e2782bb20cdfa264d1694841a7a10f63a2b361e4
for `docs.yottadb.net` pipeline job didn't work, because the if
statement didn't have quotes in it (I think). In any case, it needs to
be the same as the docker builds, because we want to build the
documentation after merge, but not for scheduled jobs.