[Pipeline] Fix readline test failures on loaded systems
All TR* tests fail when "octo" is entered in expect when waiting for
OCTO>, as octo takes time to start up, and OCTO> shows up and "octo"
entered by expect does not show up (echoing gets turned off by that
point). This doesn't really have to do with parallelization per se, but
more due to how slow the machine becomes when loaded with 8 instances of
Octo tests running concurrently all using all CPUs.

We suspect the issue is that when we start bash we don't wait for the
shell prompt, which means if bash startup turns off terminal echo
(either inside bash or a stty -echo etc. command in some system-wide or
user-specific bash startup script), it is possible the octo command will
not get echoed to the temrinal. Therefore, the right solution is to
wait for the shell prompt to show up That should fix the timing issue.