[Pipeline] Fix readline test failures on loaded systems, try 2
The previous commit (1320607cd27bdbf33fffd0a314434516ba5a0521) tried to
fix the problem, and even though it identified the proximate cause of
the problem, it was still not fixed.

It turns our that the "octo" that shows up is a rogue "octo", and
wouldn't show up with a normal user using a resonably fast system. What
happened is that the `stty cols 4096` call to change column size doesn't
finish, but expect keeps going on, thus typing "octo"; and then this
same octo shows up again later in the bash line when bash actually
processes it.

A previous solution to this was to look for `stty cols 4096`, but this
is typed out anyways, so it's basically a no-op.

The correct solution is to WAIT until #/$ shows up instead of `stty cols
4096` so that we know the stty call finished. When this happens, and we
type "octo", it is not echoed twice to the screen; but only once, in the
bash prompt.

As a result of this, all occurences of "octo" have now been removed from
the outref files.