[Pipeline] Fix re-clone error for VistA Docker pipeline job
Fixes this pipeline failure:
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/3532420077

```
git clone https://github.com/WorldVistA/docker-vista.git
fatal: destination path 'docker-vista' already exists and is not an empty directory.
```

This is similiar to this commit in the YDBDoc, and the solution is
similar: https://gitlab.com/YottaDB/DB/YDBDoc/-/commit/637a0ccfd62eb95eb1928d89206984caf9de905c.

This is a problem that only started showing up when we set-up the
pipelines locally.

The reason we have not seen this before is that since we set-up the
pipelines is that we have a nightly cron job on our gitlab runner that
deletes the old docker volumes that contain previous pipeline runs.
Hitherto, we never saw this error because we never did two merges in one
day; and so our docker-vista pipeline was "fresh" and didn't have the
previous clone. Yesterday we did, and as a result, saw this error.