[#588] Added `USER()` function to documentation
Per the discussion at https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/800#note_1228248045, the `USER()` function was not documented when !800 was merged as part of the resolution to #588. This commit adds documentation for `USER()` function that was added with that commit.

Also, the description of `CURRENT_USER` was revised for clarity and a note was added indicating that it is a synonym for `USER().