[YottaDB/DB/YDB#931][#956] Pipeline for AARCH64/ARM64 Docker Images for Octo and Octo-VistA
This is a companion commit to
https://gitlab.com/YottaDB/DB/YDB/-/commit/93a949d21606655f007b9231e35a5dfe8c8513e8,
which built YottaDB ARM64 images. Hitherto, Octo images have been
available only for AMD64 (aka x86_64).

This commit builds the docker images for Octo for AMD64 using the same
methodology done for YottaDB: Due to
https://github.com/docker/buildx/issues/1331, we could not use docker
buildx to build AARCH64 images. Instead, what we do is create the two of
them (AMD64 and AARCH64) seperately on an AMD64 and an AARCH64 machine,
tag and push them separately, and then download them back and combine
them using `docker manifest`.

This was tested against the samindcon docker repo and images have tested
on an AMD64 and an ARM64 machine.