[YottaDB/Util/YDBDevOps#9] Update pipelines to work on Ubuntu 22.04
Fix clang-tidy to version 11 and change `tools/ci/find-llvm-tool.sh` to
look up to version 15 of clang tools.

In Ubuntu 22.04, the default version for clang-tidy is 14. If we use 14,
a lot of formatting changes happen; and plus 14 is not available on
Debian 11. We chose 11 as being common in between Ubuntu 20.04, Ubuntu
22.04, and Debian 11.

`tools/ci/find-llvm-tool.sh` was not coded to find clang tool above
version 12. We now have 14 in Ubuntu 22.04; so I bumped this value up to
15.

Python as Python2 is not available on Ubuntu 22.04. Prior to commit
eefa4a6f66a05a5cfab97546738231f4952b54b4,
tests/fixtures/sqllogic/insert.py uses Python2. After discussion, we
decided to stop our auto-upgrade tests at that commit.