[YottaDB/Util/YDBDevOps#9] Undo unnecessary change to find-llvm-tool.sh in 066a5e43
`find-llvm-tool.sh` was updated in 066a5e43 to look for LLVM tools
starting from verion 15 rather than 12. That change was not necessary
but done in preparation for a future date when we move from
`clang-format-11` to `clang-format-14`.

Unfortunately, it had a side effect of picking higher versions (e.g. 14)
rather than the lower version (11), if a developer happens to have both
versions installed; and the behavior of `clang-format` for version 14
and 11 is different resulting in differently formatted `.c` files. Since
we are currently using the 11 version of the tools, we bumped the value
back down from 15 to 12, which was its original value yesterday.