[#519] Fixed failures in `TQG03` due to mishandling of `quote names` database
As reported in this thread https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1182#note_1301057700, `TQG03` has been failing the pipeline since !1182 was merged. The cause of this failure was discovered in the implementation of the new `extractQuotedColumnAlias()` label in `QueryGenerator.m`. Specifically, this label did not correctly handle the presence of spaces in double-quoted column aliases.

To resolve this issue, `extractQuotedColumnAlias()` was revised to correctly handle double-quoted column aliases. This resulted in a greatly simplified implementation that avoids the use of needless loops. The comments before and within this label definition explain the new implementation.

Since the issue is now resolved the `quotenames` database was re-enabled as a testing option in the `TQG03` test case.