[#802] Confirm drop table/recreate results with different delimiters
Prior to commit 4a1e067c, a table which was created with a specific
delimiter and then dropped (but data kept) and re-created with a new
delimiter would still use the old delimiter in its definition, resulting
in incorrect query output. This has been fixed in 4a1e067c, but a test
for that specific case has not been added. This is now done as test
`TDT08` in `test_drop_table.bats.in`.

A minor fix found while doing this: the SQL fixture for `TDT07`
incorrectly referred to itself as `TDT08`. That's also fixed now.