[YottaDB/Util/YDBDevOps#9] Update pipelines to work on Ubuntu 22.04
Octo's Dockerfile uses the RELEASED version of YottaDB. This was at
Ubuntu 20.04 until 2 days ago, when r1.38 was released, which caused a
new Docker image for the RELEASE version of YottaDB to be generated.

This commit updates the packages libicuxx and libsslxx to be the
versions available on Ubuntu 22.04; the older version are no longer
available.