[YottaDB/DB/YDBDoc#359] Encouraged use of `ydbinstall` for Octo Quickstart documentation
The Octo Quickstart documentation in `intro.rst` previously provided developer-friendly instructions for installing Octo from source. However, it would be easier for users to get up and running with Octo by using `ydbinstall.sh`. Accordingly, this commit revises `intro.rst` to encourage users to use `ydbinstall.sh` to install Octo instead of installating it from source.

This was done by replacing the instructions for installing Octo from source in `intro.rst` with the `ydbinstall.sh` instructions from the README.

In view of this change, it was also necessary to revise the README to point to the Octo developer documentation for detailed installation instructions instead of the Quickstart in `intro.rst`.

Also, renamed the `Developer's Documentation` to `Developer Documentation` (removed possessive suffix).

Additionally, prerequisites for SUSE Linux were added to `intro.rst`, and the prerequisite lists for Ubuntu and RHEL prerequisites there were alphabetized.

Finally, a comment was added to the README clarifying that `--octo` implicitly installs the dependent plugins.