[#809] Remove use of xxd/vim from Octo build to reduce extraneous dependencies
Previously, xxd was used to create a header file from the
`src/aux/octo.conf.default`, and have that be embedded inside of the
Octo binary at compile time.

On RHEL derivatives, xxd is only provided by vim; and it's weird to have
Octo depend on vim.

Removing `xxd` required us to load the config file at runtime. We
already have that code, but what we needed to find the
`src/aux/octo.conf.default` file. We already put it into the plugin
folder for Octo under $ydb_dist, but we needed to also put it in the
CMake Build `src` directory, like what we do with `ydbocto.ci` and so
the CMakeLists.txt file has been updated with an extra `configure_file`.

The code to deal with DISABLE_INSTALL in `octo_init.c` is the same as
was done for the `ydbocto.ci` file.

The auto-upgrade jobs still keep `xxd` as it's needed for previous
commits that still use `xxd`.