Fix test_psql_go_connection test failure in pipeline by installing golang from package manager
Yesterday, we got pipeline failure due to a change in lib/pq:
https://github.com/lib/pq/commit/96e73eb9aa7ba849b24eae15477456d8bbb1c9b7.

The error was as follows:
```
go/src/github.com/lib/pq/conn.go:2089:7: undefined: driver.Validator
```

After testing, it was determined that the failure is due to
`driver.Validator` not being defined in Go 1.13.8, which was hardcoded
to be used in the pipeline; as we couldn't reproduce this problem with
recent versions of go which are installed on our machines.

The solution is to remove the hardcoding of the Go version and the
manual installation and to use the Go package from the standard
repositories.