[#955] If ASAN is enabled, disable test_handle_password_message test
We have had an on-off issue for a long time with this ASAN bug:

```
 30/142 Test  #30: test_handle_password_message ............***Failed 0.02 sec
 [==========] Running 8 test(s).
 [ RUN      ] test_valid_password_message
 [       OK ] test_valid_password_message
 [ RUN      ] test_valid_input
 [       OK ] test_valid_input
 [ RUN      ] test_error_not_md5
 [       OK ] test_error_not_md5
 [ RUN      ] test_error_user_info_lookup
 [       OK ] test_error_user_info_lookup
 [ RUN      ] test_error_hash_lookup
 [       OK ] test_error_hash_lookup
 [ RUN      ] test_error_hash_conversion
 [       OK ] test_error_hash_conversion
 [ RUN      ] test_error_bad_password
 [       OK ] test_error_bad_password
 [ RUN      ] test_error_missing_username
 [       OK ] test_error_missing_username
 [==========] 8 test(s) run.
 [  PASSED  ] 8 test(s).
 Tracer caught signal 11: addr=0x5d000008 pc=0x7efdb59c77c8 sp=0x7efdae9a4d20
 ==16011==LeakSanitizer has encountered a fatal error.
 ==16011==HINT: For debugging, try setting environment variable
 LSAN_OPTIONS=verbosity=1:log_threads=1 ==16011==HINT: LeakSanitizer does not work under ptrace (strace, gdb, etc)
```

We really have never been able to figure out the ultimate cause, except
that it's caused by something in YottaDB or libICU when we are in UTF-8
mode. Actually, in this case, this is not really true, as we disabled
UTF-8 and still got the error.

`test_handle_password_message` just started crashing with LSAN recently;
it never crashed before.

We test this code everywhere else without ASAN, and that's considered
enough. For now, disable this test when ASAN is enabled.

(A small change: I noticed that a set of comments was not accurate
anymore. The commit 143fa672bcf5e88c720ed4714ed10c881163c38d referred to
disabling `test_handle_password_message`, but it was subsequently
enabled in another commit eed1b94e99da46c9c15c21deb556db8810e3115d, but
without removing the message saying that it was disabled. So the
comments saying that it is disabled are now removed.)