[#915] Run Octo bats tests concurrently (instead of serially) in pipeline
!1371 seems to have fixed most of the issues; and !1300 and !1304 fixed
some additional issues.

The changes made in this commit are:
- Make `test_errors` depend on Postgres set-up (hello_db test), as
  TERR010 requires a completed Postgres set-up, otherwise it will fail.
- Enable parallelization in build.sh using CTEST_PARALLEL_LEVEL
- Add netcat in `Dockerfile-Tests.rocky` and `Dockerfile-Tests.ubuntu`
  (missed in !1371).

The each pipeline job now seems to take around 18 minutes, compared to
52 minutes before this change.

This commit also makes a change to allow ODBC connection tests to run on
SUSE, as the name of the Postgres driver in the psqlodbc package is
"PSQL", not "Postgres".