[#915] Speed-up pipelines by pre-generating Docker Images
Testing has showed us that about 5-10 minutes of a pipeline's job is
consumed in the `apt-get install` or `yum install` step. While this is
not particularly expensive for a single job, this was repeated by and
large for all jobs.

This commit uses the new files `Dockerfile-Rocky-Tests.baseline` and
`Dockerfile-Ubuntu-Tests.baseline` to generate docker images for all
Octo test pre-requisites, and pushes them to Gitlab's Docker registry.
The push only happens on a scheduled job; so it won't happen with
regular push commits, which will use the images from the last scheduled
job (which in Octo happens nightly).

The first generation of the docker images was manually done and pushed,
but the pipeline should take care of generation when the scheduled
pipeline runs (nightly).

In the `.gitlab-ci.yml` file, the `.ubuntu` and `.rocky` sections have now
been removed.

Besides the main `Dockerfile` at the base of the repo, all other
Dockerfiles were moved to `tools/misc-dockerfile`, and the instructions
at the top of the files were updated accordingly.