[CMake] Fix errors in out of tree tests
CMake documentation recommends that your build be completely out of your
source tree. This works well in our other repos, but fails in Octo,
because there were a couple of files which hardcoded paths in assuming
that the build directory will be in the source tree, which is not a
valid assumption.

The fixes are:

- QueryGenerator.m had a `../..` path. This is fixed by putting the
  correct path in and configuring the file.
- `tests/test_regex_type_expressions.bats.in` had a `../..` path. This
  just looks like a typo, as it's correct elsewhere in the file. It was
  replaced with the correct CMake value.

After these changes, the full test suite now runs successfully out of
tree.