[#915] Fix test concurrency issue
We got this failure on the pipeline from Postgres:

```
$ cat TQG04_nocomment-037.sql.diff
1,4d0
<   Hint: No function matches the given name and argument types. You
might need to add explicit type casts.
<   Position: 74
< ERROR: function samevalue(character varying) does not exist
< SQL State: 42883
```

Investigation shows that the following tests drop/recreate the function
`samevalue` in a Postgres database:

- `tests/test_delete_from_table.bats.in` names
- `tests/test_update_table.bats.in` names
- `tests/test_values_clause.bats.in` names
- `tests/test_query_generator.bats.in` customers,pastas,names,nullcharnames,nameslastname,quotenames,northwind,boolean,nullnames,composite
- `tests/test_jdbc_connection.bats.in` customers,pastas,names,nullchar,northwind

If they run at the same time on the same database, they will cause the
above issue. This is very possible if
`tests/test_query_generator.bats.in` and
`tests/test_jdbc_connection.bats.in` run concurrently.

The solution is to move the the drop/recreate for `samevalue` to
`hello_db`, which sets up all Postgres databases. In addition, since all
databases are deleted in `hello_db`, there is no need to drop; only
create the function `samevalue`.