[YottaDB/DB/YDB#780] [V70000] Fix TYIO01 test due to journal format bump from 44 to 45
A previous commit (c4058a2815b94858713c0909fb1038e53b08e968) tried to
address this issue by bumping the journal format in the reference file.
However, we also run this test on the released version of YottaDB, which
still has the old journal format number.

The solution, implemented here, is to remove the journal format number
from the test (remove field FORMAT in table YDBJNLFRECTYPE from the
output) and just print the other fields in the table YDBJNLFRECTYPE.