[#915] Fix TQG01 failure due to concurrently running TF03 test
Below is a pipeline failure of the TQG01 subtest as described at
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/915#note_1416651730.

Postgres failed with this error:
```
> cat TQG01-099.diff
1,3c1
<                     ^
< LINE 1: DELETE FROM customersTQG01  WHERE (NOT
(samevalue(customersT...
< psql:TQG01-099.sql:1: ERROR:  relation "customerstqg01" does not exist
---
> DELETE 4
```

We suspect it's because TF03.sql (which erroenously uses the customersTQG01)
dropped that table at the time TQG01 was running.

The fix is to fix the issue with TF03 and use a test specific table name
`customersTF03`. `hello_db.bats` was also updated to drop that table
from Postgres when starting the tests, consistent with what we do with
other postgres tables.