[#979] Grammar documentation now includes a section describing the 3 different types of quotes (" ' `)
Previously, information on how to use single-quotes and double-quotes was scattered throughout the documentation in various appropriate places. However, information of backtick usage was missing and finding the information on quote usage was difficult.

To alleviate this difficulty, this commit adds a new subsection to the "Technical Notes" section describing how to use single-quotes, double-quotes, and backticks.

Additionally, it adds a new subsection head for the pre-existing documentation in the "Technical Section" for the sake of clarity and ease of reference.