[#915] Fix TMU02 test failure that happens on loaded system
See original failure here for background: https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/915#note_1438230315

Enabling parallelism as part of issue #915 (multiple commits) has caused
us to see various failures that we did not see before. In this case, `TMU02`
test failed because when Rocto got stopped, the `FORCEDHALT` message got
mixed into `rocto.log` in an unpredictable place, since rocto stopped due to
a signal delivered asynchronously.

The test failure diff looked like this:
```
Comparing outref/TMU02.ref node_usage.log
0a1
> %YDB-F-FORCEDHALT, Image HALTed by MUPIP STOP%ydboctoerror("UNKNOWNFUNCTION")=1073741834
34d34
< %ydboctoerror("UNKNOWNFUNCTION")=1073741834
```

Normally, this is not a problem, as only lines containing `%ydbocto` were kept,
but because the `FORCEDHALT` message can be anywhere, it can be mixed in with
the lines we want to keep, as shown in the failure diff.

The solution is to split the rocto log in `start_rocto` in
`test_helpers.bash` to stdout and stderr. Whether by design or by
mistake, the rocto variable dumps are in stdout, whereas all the
logging/tracing/`FORCEDHALT` is in stderr. So stderr output is what now
goes to `rocto.log`; variable dumps to go `rocto.stdout`.  `TMU02` test
updated to use `rocto.stdout`.

While redoing the TMU02 test, we noticed that grep occasionally warned
about binary data in `rocto.stdout` (previously `rocto.log`), but this
warning (itself) was filtered out along with other lines we didn't want
to see. That's a bit unclean; now a `-a` flag was added to grep to
suppress the binary data warning.