[CMake] Fix file(DOWNLOAD) usage to work on CMake < 3.17
https://gitlab.kitware.com/cmake/cmake/-/merge_requests/4159 fixes an
issue in `file(DOWNLOAD)` so that it would assume that the file needs to
be downloaded into the current directory. This is not a good assumption
in any case, so now our code has been fixed to use a full path. This
way, Octo tests can now run on CMake versions less than 3.17 (i.e.
Ubuntu 20.04).