[Pipeline] Fix Java tzdata problem
Various pipelines (e.g.
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/4734059215) failed
because of a problem in the Rocky Linux OpenJDK 11 installation that
just started happening two days ago. It was narrowed down to a broken
symlink that points to a file provided by tzdata-java. OpenJDK 17 does
not have the same problem, as it installs tzdata-java automatically. We
want to stay with Java 11 as it is still the most commonly used LTS
version.

tzdata-java is now explicitly installed since OpenJDK 11 doesn't
implicitly install it.