[#211] Fix auto-upgrade cron job failure when CREATE/DROP view is seen in TQG* tests
Issue
------

In the test-auto-upgrade job execution which failed, JDBC driver was used to run tests (the run which happens on older commit).
The output of a CREATE VIEW/DROP VIEW looked like the following
```
< # It would be non-zero only for INSERT INTO, DELETE FROM etc.
< # Note that the below would be 0 for CREATE/DROP commands
< Number of rows updated = 0
```
After this, the test-auto-upgrade job ran TQG subtest query files on the latest build of Octo which in our case lead to
running a CREATE/DROP VIEW command in the following way.

**tools/ci/build.sh**
```
1002                                 outfile="autoupgrade.$sqlfile.out"
1003                                 ../newsrc/octo -f $sqlfile > $outfile 2>&1
1004                                 ret_status=$?
```

The output generated by the above lines was not the same as the output generated by test run on the older commit. The
newer build output had CREATE VIEW/DROP VIEW instead of the JDBC driver output seen at the beginning of this comment.
Because of this the `diff` which got executed next failed.

Fix
----
* Avoid such failures by checking in `build.sh` if the query being executed in the new build is CREATE/DROP statement.
  If it is and JDBC driver was used in the older commit test execution, we skip doing `diff` on the output.
  As before, sig-11s are still validated if it happens in line 1003 above by checking ret_status value to be less than 1.

Test
----
* The existing auto-upgrade cron job and pipeline is sufficient to validate this usecase so no further addition