[#211] Fix VIEW binary definition to get auto upgraded even when only a VALUES clause is used in view definition
Issue (copied from https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1244#note_1509561251)
-----
Views which do not have any dependencies are not included in the `^%ydboctoocto("viewdependency",*)` nodes. And so, if we
encounter such views, those won't be included in the `%viewssortednames` output at all.

An example such view is the below (pasted from `TAU002_1.sql`).

```sql
create view v35 as select * from (values(1)) order by 1;
```

As part of !1426 (which converts case insensitive view names from upper case to lower case), a binary auto upgrade of the
views happened, and all views created in TAU002_1.sql did get upgraded. But `v35` got missed out. And a
`select * from v35;` which is done in `TAU002_2.sql` failed with an error that `v35` table was not found because it was
missed out in the auto upgrade (it was still named `V35` in uppercase as it did not get upgraded to lowercase name).

Fix
---
* Store the following gvn for a view when we come across any during table binary definition auto upgrade. Table definition
  by design is upgraded first so we can rely on this to ensure that the following gvn will exist during view's upgrade.
  ```
  ^%ydboctoocto("viewdependency",viewName,"values")
  ```
* Once the gvn exists `src/aux/_ydboctoViewsUpgrade.m` ensures that the view is added to %viewssortednames and thus ensuring
  auto-upgrade happens for the view.
* DROP VIEW related code is updated in src/aux/_ydboctoDiscard.m to remove this gvn when an explicit DROP VIEW is executed

Test
-----
* Additional tests are not added as test_auto_upgrade.bats.in already validates this usecase and it itself identified this
  issue during auto-upgrade cron job execution