[Pipeline] Sync with docker-vista repo update
Commit
https://github.com/WorldVistA/docker-vista/commit/e964787899ec78b787e7b350544d7049263e062e
updated the docker-vista repo to sync the settings with the
`ydb_env_set` settings (added in commit
https://gitlab.com/YottaDB/DB/YDB/-/commit/282b556908bb23d9a2417860adbb328b89e788c1)
. The most important change is that we changed the name mapping from
`%ydbAIMD` to `%ydbAIM`, and the logic in
`tests/test_vista_database.bats.in` previously relied on the presence of
`%ydbAIMD`. Now it's also changed to use `%ydbAIM`.

Historically, this change was done in AIM after the code in docker-vista
repo was written: https://gitlab.com/YottaDB/Util/YDBAIM/-/commit/c74e932c9b5f41530c6a94db67eb3c72d4600e52.
Before this commit, all AIM globals ended with "D". After this commit, we
have a new global in AIM that is `ydbAIMxref`.

The same commit in the docker-vista repo also changed AIM to be a BG
database instead of an MM database; in our case, for a temporary
pipeline run, it makes no difference, but we consistency's sake, we keep
this as BG as well.