[#1023] Bump JDBC in Octo testing from 42.6.0 to 42.7.0
There are some references in the source code for 42.6.0 which I didn't
touch since at the time of test writing 42.6.0 was the latest version;
and I don't want to manually go through and verify that indeed the same
queries are sent over in the tests.

The only change in the tests is the output of the query `show
DateStyle;`; this has been updated.

42.7.0 manually tested against Rocto using

- Squirrel SQL 4.5.1
- SQL Workbench/J Build 130
- DBeaver CE 23.1.0