[#382] Fix TDTT053 such that second difference don't cause test failure
Issue
------
The post-merge pipeline of !1533 failed in the [make-ubuntu](https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/6966844482) and the [make-tls-ubuntu](https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/6966844493) jobs. Both failures were in the `TDTT053` subtest.

```sql
$ cat TDTT053-7.sql
select cast(localtime as time);
```

```diff
$ cat TDTT053-7.diff
1c1
< 05:08:07
---
> 05:08:08
```

The reason for the failure is that we run the `localtime` function first with `MySql` and then with `Octo`. By the time the same query runs in Octo, the time has moved on to the next second and is the cause of the diff.

This is a test issue in 41f13366 that can cause random pipeline failures.

Fix
----
* The localtime query is tested separately. Same query is executed in MySql before and after the test. If diff from the test is not empty then the output of Octo
  is verfied to be within the MySql executions results which were taken before and after the test.

Misc
----
* TDTT053 was setting `mysql_is_mariadb` but this is not required by the test so the code related to it is removed