[#382] Ensure code to do IN processing for date/time types with timezone consider list item types correctly
Without the changes in current commit or 631d12b6c2a213019d1f54831c5a3652424b585e or 719a3479 following warning is seen while building
Octo in rocky linux or octo-vehu docker.
```
In file included from /tmp/octo/build/src/tmpl_print_expression.ctemplate.c:18:
/tmp/octo/build/src/tmpl_print_expression.ctemplate.c: In function 'tmpl_print_expression':
/tmp/octo/src/m_templates/template_helpers.h:33:13: warning: 'second_operand_type' may be used uninitialized in this function [-Wmaybe-uninitialized]
   written = snprintf(*global_buffer + *buffer_index, *buffer_len - *buffer_index, ##__VA_ARGS__); \
             ^~~~~~~~
/tmp/octo/build/src/tmpl_print_expression.ctemplate.c:44:35: note: 'second_operand_type' was declared here
  SqlValueType first_operand_type, second_operand_type;
                                   ^~~~~~~~~~~~~~~~~~~
```

The else block below had to be removed because it would result in second operand being treated as UKNOWN_SqlValueType when it is expected
to be the value set in line 981 for the second iteration of the do-while loop. Note second_operand_type is only used when
second_operand_needs_timezone_comp is TRUE. When this happens second_operand_needs_timezone_check is set to FALSE such that second
iteration of the do-while loop would carry forward the value set at line 981 for the rest of the iterations. This optimization was done
to prevent repeated invocation of `lp_get_plan_value_type();` for each item in IN list as we know that all items in the IN list is considered
as same type. But with the else block, second_operand_type
would reset the value set by line 981 and this resulted in TDTT081 failure.
```
 971                         /* All items in the IN list is of same type but its possible that we see NULL values in between so check
 972                          * for TIME ZONE type value inside the loop. Once found, the search will not happen again and NULL values
 973                          * also go through Conversion logic but the routine invoked returns immediately if NULL value is seen so
 974                          * this should be okay.
 975                          */
 976                         // Prevent date/time type search if the first operand is not of similar type
 977                         second_operand_type = UNKNOWN_SqlValueType; // avoids false [-Wmaybe-uninitialized] warning
 978                         boolean_t second_operand_needs_timezone_check = (IS_DATE_TIME_TYPE(first_operand_type));
 979                         do {
 980                                 if (second_operand_needs_timezone_check) {
 981                                         second_operand_type = lp_get_plan_value_type(plan2);
 982                                         if ((TIMESTAMP_WITH_TIME_ZONE_LITERAL == second_operand_type)
 983                                             || (TIME_WITH_TIME_ZONE_LITERAL == second_operand_type)) {
 984                                                 second_operand_needs_timezone_comp = TRUE;
 985                                                 // re-use the below variable to prevent repeated search
 986                                                 second_operand_needs_timezone_check = FALSE;
 987                                         }
-				     } else {
-					second_operand_type = UNKNOWN_SqlValueType; // avoids false [-Wmaybe-uninitialized] warning
				     }
```
To avoid the compilation warning and retain the optimization,
initialization code for second_operand_type is done in line 977 and the else block is removed.

Following is the type of query which failed prior to this commit.
```sql
create table TDTT005timetz (id integer, dob time with time zone );
insert into TDTT005timetz values(1,time with time zone'01:01:01+05:00');
select time'01:01:00' NOT IN (n1.dob,time'01:01:00') from TDTT005timetz n1;
```
result of the select is expected to be 1 but it was 0 prior to this commit.

With this commit the result is as expected. This is tested by TDTT081.