[#957] Use `WARN_FEATURE_NOT_IMPLEMENTED` to issue a warning for usage of features that are allowed but not fully implemented
`WARN_FEATURE_NOT_IMPLEMENTED` is added in this commit to be used wherever Octo is allowing the usage of
some features which are not fully implemented. Octo allows such usages to enable few clients. Refer to
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/957#note_1943550507 for more information. Previous to
this change `ERR_FEATURE_NOT_IMPLEMENTED` was being used even for this case.

`ERR_FEATURE_NOT_IMPLEMENTED` still exists and is used wherever an ERROR is issued for features that are not implemented.

`WARN_FEATURE_NOT_IMPLEMENTED` and `ERR_FEATURE_NOT_IMPLEMENTED` have the same text and this resulted in
pre-commit hook giving the following error. This was ignored by using `-n` while committing the changes as it didn't
seem like the text description needed to change for the new message.
```
WARN_FEATURE_NOT_IMPLEMENTED
ERR_FEATURE_NOT_IMPLEMENTED

  --> Hook .git/hooks/pre-commit returned non-zero status
  --> HOOK-E-FAIL : Non-zero exit status returned by tools/ci/doc_error_update.sh
```

Test changes done are to update the ERR_FEATURE_NOT_IMPLEMENTED references to WARN_FEATURE_NOT_IMPLEMENTED.