[#382] Fix run_query_and_crosscheck_avoiding_execution_time_delay to not remove space between date and time
Issue
-------
Master branch pipeline(https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/7102102436) failed for
31a02352785374843c5db096f95ebe7f69c34cd7 because TDTT053 failed with the following output.
```
date: invalid date ‘2024-06-1409:39:06’
```

Fix
------
TDTT053 invokes run_query_and_crosscheck_avoiding_execution_time_delay to do testing.
When "mysql" is the third argument the routine is expected to not remove any space from the query result.
31a02352785374843c5db096f95ebe7f69c34cd7 took care of fixing the first `if` condition and the code in it
to do the filtering only when "postgres" is passed and if space is found at the beginning of the output.
But, it forgot about the second `if` condition which is executed when there is a difference
in MySql/Postgres and Octo result due to time of execution delay. This is fixed by this commit as shown below.
```diff
@@ -1467,8 +1467,8 @@ run_query_and_crosscheck_avoiding_execution_time_delay() {
-			if [[ $ispostgres ]]; then
-				sed -i 's/ \(.*\)/\1/g' "$name"_output2.txt
+			if [[ $ispostgres -eq 1 ]]; then
+				sed -i 's/^ \(.*\)/\1/g' "$name"_output2.txt
```