Fix commit used by tools/ci/build.sh to do boolean result column transformation from 1 to t and 0 to f
Issue
-------
@nars1 reported that test-auto-upgrade cron job failed for the latest changes.
More description of the failure is in the **Analysis** section. This needs to be fixed.

Analysis
----------
Old commit selected  - 988484b7

Diff of a failed test shows the following:
```diff
< 0
< 0
< 0
< 0
---
> f
> f
> f
> f
```

This was supposed to be fixed by https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1581 tools/ci/build.sh changes
by transforming f to 0.

Following is the description of the change from !1581

    * test-auto-upgrade job started failing with this change because if older commit is selected as or before
      584ac8c50da72878ecf744cd6ee9ed7806a89996 (one commit before this commit) then result table will have 1 and 0 for t and f.
      This caused tools/ci/build.sh to fail when comparing TQG query results of old build with new build. This is fixed by converting
      new build test result to have 1 and 0 instead of t and f by applying a `sed` command before the comparison. Changes are in
      tools/ci/build.sh. This fixed the issue.

Above described change makes use of commit 584ac8c50da72878ecf744cd6ee9ed7806a89996 to know if the transformation is needed. The change specifies
that this is one commit before 63a8f89903748837ebdd875734075369e24737f2 (!1581). But the commit history shown below shows
that commit 2 is the actual commit just before the fix. This must have been merged before !1581 got merged.
1) 63a8f89903748837ebdd875734075369e24737f2
2) 988484b7cff7859eec77d653ce693ef8a43b5d3b
3) 584ac8c50da72878ecf744cd6ee9ed7806a89996

More details on how the test failed:
* Before 1st commit Octo outputs boolean values as 1 and 0. After the commit t and f is the output.
test_auto_upgrade test compares result of old build and new build for TQG tests.
Because of the change in output behavior, 1st commit modified tools/ci/build.sh to transform the result of
new build from having t and f to 1 and 0. The change makes use of 3rd commit to check if this transformation
is needed. This is the issue. In the failed cron job the commit selected is 2. This commit must have been
merged before 1 got merged causing the issue.

Changing tools/ci/build.sh to make use of commit 2 instead of 3 to do the transformation will fix this issue.

Changes
--------
* tools/ci/build.sh changes diff is shown below
```diff
--- a/tools/ci/build.sh
+++ b/tools/ci/build.sh
@@ -1290,7 +1290,7 @@ else
                                                fi
                                                mv $outfile.tmp $outfile
                                        fi
-                                       pre_octo1032_commit="584ac8c50da72878ecf744cd6ee9ed7806a89996" # 1 commit before #1032 commit
+                                       pre_octo1032_commit="988484b7cff7859eec77d653ce693ef8a43b5d3b" # 1 commit before #1032 commit
                                        # Disable the "set -e" setting temporarily as the "git merge-base" can return exit status 0 or 1
                                        set +e
                                        git merge-base --is-ancestor $commitsha $pre_octo1032_commit
@@ -1299,7 +1299,7 @@ else
                                        set -e
                                        if [[ (0 == $is_post_octo1032_commit) && (("select" == $querytype) || ("values" == $querytype)) ]]; then
                                                # Octo before #1032 returned boolean values as 1 or 0. So, if the old build version is or before
-                                               # 584ac8c50da72878ecf744cd6ee9ed7806a89996 (1 commit before #1032 change) convert boolean values of
+                                               # 988484b7cff7859eec77d653ce693ef8a43b5d3b (1 commit before #1032 change) convert boolean values of
                                                # the new build such that it can be compared with the old results.
                                                # Converts t to 1 and f to 0. The tranformation is applied twice to convert
                                                # consecutive values like tt to 11.

```
* This fixed the issue. Confirmed by running test-auto-upgrade job by selecting old commit as the 2nd commit.